backend/requirements.txt
# PDF and Image Processing
pdf2image==1.17.0
# pytesseract removed in favor of easyocr
easyocr
numpy
Pillow

# LLM and API handling
google-generativeai==0.4.1
python-dotenv==1.0.1

# Database
pymongo==4.6.1
dnspython==2.6.1

# Web API
fastapi
uvicorn[standard]

# Authentication
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1

# HTTP Client (for reverse geocoding)
httpx

# Data Validation
pydantic[email]